PG100806A: this was supposed to be a revolution in strands, but isn't much after all. Read RTAGS_PG.

+++ Apps/SFU/HEART_MONITOR_EXPERIMENTAL/app.c
@@ -909,7 +909,7 @@ endthread
 #define	LI_GETCMD	2
 #define	LI_PERSTAT	3
 
-#define	MYFD	((int)data)
+#define	MYFD	data
 
 strand (listener, int)
